Fix bug with being unable to craft fluid if fluid was present in output, make extract work

+++ common/src/main/java/xyz/mackan/crystallurgy/recipe/FluidSynthesizerRecipe.java
@@ -72,7 +72,7 @@ public class FluidSynthesizerRecipe implements Recipe<SimpleInventory> {
         return list;
     }
 
-    public boolean matchFluid(World world, Fluid checkFluid, int checkAmount) {
+    public boolean matchFluid(World world, Fluid checkFluid, long checkAmount) {
         // inputFluid is from the recipe.
         if (world.isClient()) {
             return false;
